Refactor API tickets   - Add test for tickets API   - Refactor API tickets   - ADD constraint foreign key to table

+++ module/ticket.js
@@ -3,36 +3,55 @@ const pool = require('./db');
 async function issueNewTicket (req, res) {
   try {
     const userId = Number(req.body.userId);
+    if( isNaN(userId) ) return res.status(400).send("Error wrong parameter");
+
     const client = await pool.connect();
+    const userInfo = await client.query(`SELECT * FROM t_user WHERE id = $1`, [userId]);
+
+    if( userInfo.rowCount === 0 ) {
+      return res.status(404).send('Error no User');
+    }
+
     const result = await client.query(`
     WITH newTicket AS (
-      INSERT INTO t_ticket(ownerId) VALUES($1) RETURNING *
+      INSERT INTO t_ticket(owner_id) VALUES($1) RETURNING *
     )
-    SELECT tnt.id as ticketId, tu.id as userId, tu.name as userName, tnt.issuetime
+    SELECT tnt.id as "ticketId", 
+           tu.id as "userId", 
+           tu.name as "userName", 
+           tnt.issue_time as "issueTime"
       FROM newTicket as tnt, t_user as tu
-     WHERE tnt.ownerId = tu.id;
+     WHERE tnt.owner_id = tu.id;
     `, [userId]);
     client.release();
-    return res.status(200).json({
+
+    if( result.rowCount === 0 ) return res.status(404).end('Error wrong user');
+    
+    return res.status(201).json({
       rows:result.rows,
       rowCount: result.rowCount
     })
   } catch (err) {
-    console.error(err);
-    res.send("Error " + err);
+    // console.error(err);
+    res.status(500).end("Error " + err);
   }
 }
 async function deleteTicket (req, res) {
   try {
     const ticketId = Number(req.params.ticketId);
+    if( isNaN(ticketId) ) return res.status(400).send("Error wrong parameter");
+
     const client = await pool.connect();
     const result = await client.query(`
     WITH newTicket AS (
       DELETE FROM t_ticket WHERE id=$1 RETURNING *
     )
-    SELECT tnt.id as ticketId, tu.id as userId, tu.name as userName, tnt.issuetime
+    SELECT tnt.id as "ticketId", 
+           tu.id as "userId", 
+           tu.name as "userName", 
+           tnt.issue_time as "issueTime"
       FROM newTicket as tnt, t_user as tu
-     WHERE tnt.ownerId = tu.id;
+     WHERE tnt.owner_id = tu.id;
     `, [ticketId]);
     client.release();
     return res.status(200).json({
@@ -48,10 +67,13 @@ async function getTicketList (req, res) {
   try {
     const client = await pool.connect();
     const result = await client.query( `
-    SELECT tt.id as ticketId, tt.ownerid, tu.name as userName, tt.issueTime
+    SELECT tt.id as "ticketId", 
+           tt.owner_id as "ownerId", 
+           tu.name as "userName", 
+           tt.issue_time as "issueTime"
       FROM t_ticket as tt
       LEFT OUTER JOIN t_user as tu
-      ON tu.id = tt.ownerid;
+      ON tu.id = tt.owner_id;
     `);
     client.release();
     return res.status(200).json({
@@ -68,10 +90,13 @@ async function getTicketListByUser (req, res) {
     let userId = Number(req.params.userId);
     const client = await pool.connect();
     const result = await client.query( `
-    SELECT tt.id as ticketId, tt.ownerid, tu.name as userName, tt.issueTime
+    SELECT tt.id as ticketId, 
+           tt.owner_id, 
+           tu.name as userName, 
+           tt.issue_time
       FROM t_ticket as tt
       LEFT OUTER JOIN t_user as tu
-      ON tu.id = tt.ownerid
+      ON tu.id = tt.owner_id
      WHERE tu.id = $1;
     `, [userId]);
     client.release();
@@ -80,28 +105,36 @@ async function getTicketListByUser (req, res) {
       rowCount: result.rowCount
     })
   } catch (err) {
-    console.error(err);
+    console.error('getTicketListByUser', err);
     res.send("Error " + err);
   }
 }
 async function getTicketDetail (req, res) {
   try {
     const id = Number(req.params.ticketId);
+    if( isNaN(id) ) return res.status(400).send("Error wrong parameter");
+
     const client = await pool.connect();
     const result = await client.query( `
-    SELECT tt.id as ticketId, tt.ownerid, tu.name as userName, tt.issueTime
+    SELECT tt.id as "ticketId", 
+           tt.owner_id "ownerId", 
+           tu.name as "userName", 
+           tt.issue_time as "issueTime"
       FROM t_ticket as tt
       LEFT OUTER JOIN t_user as tu
-      ON tu.id = tt.ownerid
+      ON tu.id = tt.owner_id
       WHERE tt.id = $1;
     `, [id]);
     client.release();
+
+    if (result.rowCount === 0 ) return res.status(404).end('Error No Ticket Data');
+    
     return res.status(200).json({
       rows:result.rows,
       rowCount: result.rowCount
     })
   } catch (err) {
-    console.error(err);
+    console.error('getTicketDetail', err);
     res.send("Error " + err);
   }
 }
@@ -117,7 +150,7 @@ async function deleteTicketList (req, res) {
       rowCount: result.rowCount
     })
   } catch (err) {
-    console.error(err);
+    console.error('deleteTicketList', err);
     res.send("Error " + err);
   }
 }
